Adds request classes
